requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
joblib
psycopg2-binary
python-dotenv
mlflow==2.14.2
# Pin scikit-learn to the version used when training the model
scikit-learn==1.5.1
pika==1.3.2
